+++ ols/utils/ssl.py
@@ -1,7 +1,5 @@
 """Utility function for retrieving SSL version and list of ciphers for TLS secutiry profile."""
 
-# ruff: noqa:A005
-
 import logging
 from typing import Optional
