+++ pyproject.toml
@@ -13,10 +13,6 @@ target-version = "py311"
 lint.pydocstyle.convention = "google"
 line-length = 100
 
-[tool.pdm.version]
-source = "file"
-path = "ols/version.py"
-
 [tool.coverage.report]
 # unit tests fails if the total coverage measurement is under this threshold value
 fail_under = 90
@@ -37,45 +33,27 @@ exclude_dirs = ["tests", "scripts"]
 
 [tool.pdm]
 distribution = true
+[tool.pdm.version]
+source = "file"
+path = "ols/version.py"
 
-[tool.pdm.dev-dependencies]
-dev = [
-    "black==25.1.0",
-    "pydocstyle==6.3.0",
-    "fastparquet==2024.5.0",  # Required for model evaluation (runtime, if parquet qna file is used)
-    "httpx==0.27.2",
-    "mypy==1.14.1",  # Usually needs to be set to latest stable version available
-    "packaging==24.1",
-    "pytest==8.3.2",
-    "pytest-cov==5.0.0",
-    "pytest-asyncio==0.24.0",
-    "pydantic==2.9.2",
-    "rouge-score==0.1.2",  # Required for model evaluation
-    "ruff==0.9.0",
-    "bandit==1.7.9",
-    "types-requests==2.32.0.20240622",
-    "gradio==4.44.1",
-    "boto3==1.34.145",
-    "pyroscope-io==0.8.8",
-    "memray==1.15.0",
-    "reportportal-client==5.5.6",
-    "pytest-reportportal==5.4.1",
-    "pytest-benchmark[histogram]>=4.0.0",
-    "typing-extensions==4.12.2",
-    "pytest-subtests==0.13.1",
-    "build==1.2.2.post1",
-    "twine==5.1.1",
-    "pylint==3.3.2",
-]
-
-# The following section is needed only for torch[cpu] variant on Linux,
-# because this variant need to be downloaded from external link, not from
-# standard Python package index:
-#
-# [[tool.pdm.source]]
-# type = "find_links"
-# url = "https://download.pytorch.org/whl/cpu/torch_stable.html"
-# name = "torch"
+[tool.pdm.scripts]
+start = "pdm run make run"
+test = "pdm run make test"
+test-unit = "pdm run make test-unit"
+test-integration = "pdm run make test-integration"
+test-e2e = "pdm run make test-e2e"
+coverage-report = "pdm run make coverage-report"
+unit-tests-coverage-report = "pdm run make unit-tests-coverage-report"
+integration-tests-coverage-report = "pdm run make integration-tests-coverage-report"
+check-types = "pdm run make check-types"
+generate-schema = "pdm run make schema"
+security-check = "pdm run make security-check"
+benchmarks = "pdm run make benchmarks"
+requirements = "pdm run make requirements.txt"
+version = "pdm run make print-version"
+verify-sources = "pdm run make verify"
+verify-packages = " pdm run make verify-packages-completeness"
 
 
 [build-system]
@@ -88,7 +66,7 @@ dynamic = ["version"]
 description = "Road-core service is an AI powered assistant that runs on OpenShift and provides answers to product questions using backend LLM services."
 authors = []
 dependencies = [
-    "pdm==2.21.0",
+    "pdm>=2.21.0",
     'torch@http://download.pytorch.org/whl/cpu/torch-2.5.1%2Bcpu-cp311-cp311-linux_x86_64.whl ; platform_system != "Darwin"',
     'torch-macos@http://download.pytorch.org/whl/cpu/torch-2.5.1-cp311-none-macosx_11_0_arm64.whl ; platform_system == "Darwin"',
     "pandas>=2.2.3",
@@ -145,27 +123,39 @@ license = {file = "LICENSE"}
 Homepage = "https://github.com/road-core/service"
 Issues = "https://github.com/road-core/service/issues"
 
-[tool.pdm.scripts]
-start = "pdm run make run"
-test = "pdm run make test"
-test-unit = "pdm run make test-unit"
-test-integration = "pdm run make test-integration"
-test-e2e = "pdm run make test-e2e"
-coverage-report = "pdm run make coverage-report"
-unit-tests-coverage-report = "pdm run make unit-tests-coverage-report"
-integration-tests-coverage-report = "pdm run make integration-tests-coverage-report"
-check-types = "pdm run make check-types"
-generate-schema = "pdm run make schema"
-security-check = "pdm run make security-check"
-benchmarks = "pdm run make benchmarks"
-requirements = "pdm run make requirements.txt"
-version = "pdm run make print-version"
-verify-sources = "pdm run make verify"
-verify-packages = " pdm run make verify-packages-completeness"
-
 [tool.setuptools]
 packages = ["ols"]
 
 [tool.pylint."MESSAGES CONTROL"]
 good-names = ["e"]
 disable = ["C0301", "C0302", "E0602", "E0611", "E1101", "R0902", "R0903", "R0913", "R0914", "W0102", "W0212", "W0511", "W0613", "W0621", "W0707", "W0718", "W0719", "R0801", "R0917"]
+
+[dependency-groups]
+dev = [
+    "black==25.1.0",
+    "pydocstyle==6.3.0",
+    "fastparquet==2024.5.0",
+    "httpx==0.27.2",
+    "mypy==1.14.1",
+    "packaging==24.1",
+    "pytest==8.3.2",
+    "pytest-cov==5.0.0",
+    "pytest-asyncio==0.24.0",
+    "pydantic==2.9.2",
+    "rouge-score==0.1.2",
+    "ruff>=0.11.7",
+    "bandit==1.7.9",
+    "types-requests==2.32.0.20240622",
+    "gradio==4.44.1",
+    "boto3==1.34.145",
+    "pyroscope-io==0.8.8",
+    "memray==1.15.0",
+    "reportportal-client==5.5.6",
+    "pytest-reportportal==5.4.1",
+    "pytest-benchmark[histogram]>=4.0.0",
+    "typing-extensions==4.12.2",
+    "pytest-subtests==0.13.1",
+    "build==1.2.2.post1",
+    "twine==5.1.1",
+    "pylint==3.3.2",
+]
