+++ scripts/generate_openapi_schema.py
@@ -50,7 +50,7 @@ def read_version_from_pyproject():
     # it is not safe to just try to read version from pyproject.toml file directly
     # the PDM tool itself is able to retrieve the version, even if the version
     # is generated dynamically
-    completed = subprocess.run(  # noqa: S603
+    completed = subprocess.run(
         ["pdm", "show", "--version"],  # noqa: S607
         capture_output=True,
         check=True,
