+++ tests/integration/test_project_version_consistency.py
@@ -25,7 +25,7 @@ def read_version_from_pyproject():
     # it is not safe to just try to read version from pyproject.toml file directly
     # the PDM tool itself is able to retrieve the version, even if the version
     # is generated dynamically
-    completed = subprocess.run(  # noqa: S603
+    completed = subprocess.run(
         ["pdm", "show", "--version"], capture_output=True, check=True  # noqa: S607
     )
     return completed.stdout.decode("utf-8").strip()
